initial recent files function
created an initial function that adds recently generated lattice files
under the files>recent files menu. There is a bug that opens the file
automatically, but also gets an error where the file directory is a
none type. have commented out the line of code that automatically opens
the file