Slowly refactoring to account for the fact that LOUT flag specifies the output so a user could specify more or less output. This requires that the parser be intelligent enough to figure out what data it is reading. More work is required here, but I am committing at a place where we can restore previous functionality if something goes afoul.
Also adding a Tesla Test Facility example. I am using this instead of VISA because TTF has segmented undulators and a more complicated lattice. Numerical tests show that ttf-base.lat and ttf-base.out reproduce the output of the TTF examples downloadable from the Genesis website.

Also adding a simple class which executes Genesis executable from the command line for support of execution in RadTrack. This is a very lightweight class, and could probably be expanded for further functionality later.