rbExecuteGenesis now executes, although it uses os.system instead of the subprocess library. I note the Python 3 documentation here:
"The subprocess module provides more powerful facilities for spawning new processes and retrieving their results; using that module is preferable to using this function. See the Replacing Older Functions with the subprocess Module section in the subprocess documentation for some helpful recipes."

However, I note that I could not get the subprocess library to run Genesis, so for now we will stick with os.system.