Old unit tests in proper format
Some of the tests can't (at the moment) be separated from the Qt
infrastructure. So, there has to be an instance of QApplication running.
Since importExport_test.py is the first test run to require
QApplication, it is instantiated here. Previous crashes at the end of
running the tests was due to more than one instance of QApplication
running.