renamed top-level examples directory
'examples' was misleading, so I changed the name to 'experimental'
This code was intended in part as a sandbox for testing out the RadTrack
API and in part has a place to develop tests of individual RadTrack
modules.
Some of this code should be used to develop proper tests.
Some of this code will be useful for further developing native RadTrack
features.