Copy imported files to sessionDirectory
Imported files, as well as referenced files, get copied to the
sessionDirectory so simulations can run. See issue #82. Needs testing
with working files.

For now, every file that has the same name as an element parameter is
copied. This is because every different kind of element has a different
parameter name for associated files (INPUTFILE, WAVEFORM, FIELDMAP,
etc.)