Dummy import/export tab methods
This allows the try/except code in RbGlobal to be removed. Errors should
be handled at the top level and either logged or presented to the user
for action.

LaserTab contains invalid code from BunchTab, so it exits early after
creating the file, allowing it to be reloaded in projects (even without
saved data).