Fix bunch tab to allow saving
If the user has entered erroneous data in the Bunch tab before closing
the application, these data are replaced to allow saving the tab without
error messages stopping the process. If the user is closing the
application, that is not the time to ask them to do more work to fix
mistakes.