Handle beam creation and importation in Bunch Tab
The Bunch Tab now has two separate modes of operation: creating a beam
bunch with the twiss parameter tables and importing a beam bunch from a
file. In the former case, the user can edit any table and plots are
updated automatically (negating the need for a Generate Bunch button).
In the latter case, all controls that affect the beam distribution are
disabled so as to not lose the user's data. The number of particles, the
momentum, and the charge are kept enabled since these don't affect the
distribution of particles. If the number of particles specified is less
than the number imported, then a random sample is taken from the
original bunch.

Additional code distinguishes between tabs with created data and
imported data on reloading a project.

In order to speed up the user's work flow, plots only work with a
maximum of 10,000 particles.