Allow user to input Elegant simulation files
When a user inputs an Elegant simulation file, the Elegant simulation tab loads it into a text editor where it can be viewed, edited, and run. The default Elegant simulation tab still has the combo boxes for selecting parameters in other tabs. There is now a button to convert these menu choices into an Elegant file where it can be edited for more advanced simulations.